+++ packages/orq-rc/src/models/operations/updatedataset.ts
@@ -182,7 +182,7 @@ export const UpdateDatasetResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-04-21T12:47:47.709Z",
+    "2026-04-21T14:45:57.733Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {