+++ packages/orq-rc/RELEASES.md
@@ -8479,3 +8479,13 @@ Based on:
 - [typescript v4.8.0-rc.40] packages/orq-rc
 ### Releases
 - [NPM v4.8.0-rc.40] https://www.npmjs.com/package/@orq-ai/node/v/4.8.0-rc.40 - packages/orq-rc
+
+## 2026-04-21 14:49:08
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.761.9 (2.881.4) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.8.0-rc.42] packages/orq-rc
+### Releases
+- [NPM v4.8.0-rc.42] https://www.npmjs.com/package/@orq-ai/node/v/4.8.0-rc.42 - packages/orq-rc